# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Exactamente API",
    "description": "Recursos universitarios de Exactas: parciales, resúmenes y finales.\n\nEste documento se genera desde los schemas Zod del backend, así que no\npuede desincronizarse de la validación. Los clientes (web, panel admin y\nservidor MCP) generan sus tipos desde acá.\n\n**Paginación:** `page` (default 1) y `limit` (default 20, máx 100).\nExcepción: `GET /careers` devuelve `{ data: [...] }` sin paginar.\n\n**Auth:** las rutas protegidas piden `Authorization: Bearer <jwt>`.\nLas de `/admin/*` requieren rol `admin` o `superadmin`.",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://github.com/exactamente-ar/exactamente-backend"
    }
  },
  "servers": [
    {
      "url": "https://api.exactamente.com.ar/api/v1",
      "description": "Producción"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "bearer",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ResourceType": {
        "id": "ResourceType",
        "type": "string",
        "enum": [
          "resumen",
          "parcial",
          "final"
        ]
      },
      "ResourceSubtype": {
        "id": "ResourceSubtype",
        "type": "string",
        "enum": [
          "parcial",
          "recuperatorio",
          "prefinal",
          "parcialito"
        ]
      },
      "ResourceStatus": {
        "id": "ResourceStatus",
        "type": "string",
        "enum": [
          "pending",
          "published",
          "rejected"
        ]
      },
      "Resource": {
        "id": "Resource",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "subjectId": {
            "type": "string"
          },
          "title": {
            "description": "Para resúmenes lo pone el usuario; para parcial/final se auto-genera",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ResourceType"
          },
          "subtype": {
            "description": "Solo para type=parcial",
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceSubtype"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ResourceStatus"
          },
          "examYear": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "examMonth": {
            "description": "1–12",
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "examDay": {
            "description": "1–31; requerido para type=final",
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "topic": {
            "description": "Tema, 1–5",
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "downloadCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "publishedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "fileUrl": {
            "description": "URL pública ya resuelta. No construir URLs a mano; null si no está publicado",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "subjectId",
          "title",
          "type",
          "subtype",
          "status",
          "examYear",
          "examMonth",
          "examDay",
          "topic",
          "notes",
          "downloadCount",
          "publishedAt",
          "createdAt",
          "fileUrl"
        ],
        "additionalProperties": false
      },
...
    }
  },
...
  "paths": {
...
    "/api/v1/resources": {
      "get": {
        "operationId": "getApiV1Resources",
        "tags": [
          "Recursos"
        ],
        "summary": "Listar recursos publicados",
        "description": "Solo devuelve `status: published`, ordenados por `publishedAt` descendente.",
        "responses": {
          "200": {
            "description": "Recursos publicados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Resource"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "page": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "totalPages": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "totalPages"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "postApiV1Resources",
        "tags": [
          "Recursos"
        ],
        "summary": "Subir un recurso",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "multipart/form-data. Solo PDF, máximo 20 MB. Queda en `status: pending` hasta que un admin lo apruebe.",
        "responses": {
          "201": {
            "description": "Recurso creado, pendiente de moderación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subjectId": {
                      "type": "string"
                    },
                    "title": {
                      "description": "Para resúmenes lo pone el usuario; para parcial/final se auto-genera",
                      "type": "string"
                    },
                    "type": {
                      "$ref": "#/components/schemas/ResourceType"
                    },
                    "subtype": {
                      "description": "Solo para type=parcial",
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/ResourceSubtype"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "status": {
                      "$ref": "#/components/schemas/ResourceStatus"
                    },
                    "examYear": {
                      "anyOf": [
                        {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "examMonth": {
                      "description": "1–12",
                      "anyOf": [
                        {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "examDay": {
                      "description": "1–31; requerido para type=final",
                      "anyOf": [
                        {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "topic": {
                      "description": "Tema, 1–5",
                      "anyOf": [
                        {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "notes": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "downloadCount": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "publishedAt": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    },
                    "fileUrl": {
                      "description": "URL pública ya resuelta. No construir URLs a mano; null si no está publicado",
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "subjectId",
                    "title",
                    "type",
                    "subtype",
                    "status",
                    "examYear",
                    "examMonth",
                    "examDay",
                    "topic",
                    "notes",
                    "downloadCount",
                    "publishedAt",
                    "createdAt",
                    "fileUrl"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Sin token o token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limit excedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/resources": {
      "get": {
        "operationId": "getApiV1AdminResources",
        "tags": [
          "Admin"
        ],
        "summary": "Listar recursos en cualquier estado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Recursos, filtrables por status, subjectId y careerId",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminResource"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "page": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "totalPages": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "totalPages"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Sin token o token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "postApiV1AdminResources",
        "tags": [
          "Admin"
        ],
        "summary": "Subir recurso como admin",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Publica directo salvo que se mande `autoPublish=false`, que lo deja pending.",
        "responses": {
          "201": {
            "description": "Recurso creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subjectId": {
                      "type": "string"
                    },
                    "title": {
                      "description": "Para resúmenes lo pone el usuario; para parcial/final se auto-genera",
                      "type": "string"
                    },
                    "type": {
                      "$ref": "#/components/schemas/ResourceType"
                    },
                    "subtype": {
                      "description": "Solo para type=parcial",
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/ResourceSubtype"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "status": {
                      "$ref": "#/components/schemas/ResourceStatus"
                    },
                    "examYear": {
                      "anyOf": [
                        {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "examMonth": {
                      "description": "1–12",
                      "anyOf": [
                        {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "examDay": {
                      "description": "1–31; requerido para type=final",
                      "anyOf": [
                        {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "topic": {
                      "description": "Tema, 1–5",
                      "anyOf": [
                        {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "notes": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "downloadCount": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "publishedAt": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    },
                    "fileUrl": {
                      "description": "URL pública ya resuelta. No construir URLs a mano; null si no está publicado",
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "subjectTitle": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "careerName": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "planName": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "planYear": {
                      "anyOf": [
                        {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "uploadedBy": {
                      "description": "ID del usuario que lo subió",
                      "type": "string"
                    },
                    "reviewedBy": {
                      "description": "ID del admin que lo moderó",
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "rejectionReason": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    }
                  },
                  "required": [
                    "id",
                    "subjectId",
                    "title",
                    "type",
                    "subtype",
                    "status",
                    "examYear",
                    "examMonth",
                    "examDay",
                    "topic",
                    "notes",
                    "downloadCount",
                    "publishedAt",
                    "createdAt",
                    "fileUrl",
                    "subjectTitle",
                    "careerName",
                    "planName",
                    "planYear",
                    "uploadedBy",
                    "reviewedBy",
                    "rejectionReason",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Sin token o token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}